email_service/requirements.txt
flask
Flask-Smorest
Flask-Mail
marshmallow
python-dotenv
gunicorn
redis
pika